adds UI
